Warning fixes

+++ alsfvm/src/reconstruction/ENOCUDA.cu
@@ -246,9 +246,9 @@ void ENOCUDA<Equation, order>::computeDividedDifferences(const memory::Memory<re
     const int nz = input.getSizeZ();
 
     // Sanity check, we need at least ONE point in the interior.
-    assert(nx > 2*direction.x * level);
-    assert(ny > 2*direction.y * level);
-    assert(nz > 2*direction.z * level);
+    assert(nx > int(2*direction.x * level));
+    assert(ny > int(2*direction.y * level));
+    assert(nz > int(2*direction.z * level));
 
     const ivec3 start = int(level) * direction;
     const ivec3 end = ivec3(nx, ny, nz) - int(level) * direction;
